requirements.txt
# ===================================================================
# XGif Dependencies
# ===================================================================

# UI Framework
wxPython>=4.2.0,<5.0.0

# Screen Capture
dxcam>=0.0.5,<1.0.0  # DXGI Desktop Duplication (optional high-perf backend)

# Image Processing
numpy>=2.0.0,<3.0.0
Pillow>=10.0.0,<11.0.0
imageio>=2.33.0,<3.0.0
# imageio-ffmpeg removed (uses system FFmpeg via IMAGEIO_FFMPEG_EXE)

# GPU Monitoring
nvidia-ml-py>=12.0.0,<13.0.0  # pynvml

# System Monitoring
psutil>=5.9.0,<6.0.0

# Audio Recording
sounddevice>=0.4.6,<1.0.0
soundfile>=0.12.0,<1.0.0
# pydub removed (uses FFmpeg subprocess instead)

# Keyboard Input Detection
pynput>=1.7.6,<2.0.0

# ===================================================================
# Optional (not in requirements, install manually if needed)
# ===================================================================
# GPU Acceleration (NVIDIA):
#   pip install "cupy-cuda12x>=12.0.0"
# Audio mixing without FFmpeg:
#   pip install "scipy>=1.10.0,<2.0.0"
